+++ previous-versions/accounting/RELEASES.md
@@ -709,3 +709,13 @@ Based on:
 - [typescript v0.38.2] previous-versions/accounting
 ### Releases
 - [NPM v0.38.2] https://www.npmjs.com/package/@codat/accounting/v/0.38.2 - previous-versions/accounting
+
+## 2023-09-29 00:03:54
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.39.0] previous-versions/accounting
+### Releases
+- [NPM v0.39.0] https://www.npmjs.com/package/@codat/accounting/v/0.39.0 - previous-versions/accounting